+++ netty-loom-spring-core/src/main/java/io/github/azholdaspaev/nettyloomspring/core/pipeline/NettyPipelineConfigurer.java
@@ -1,8 +1,0 @@
-package io.github.azholdaspaev.nettyloomspring.core.pipeline;
-
-import io.netty.channel.ChannelPipeline;
-
-public interface NettyPipelineConfigurer {
-
-    void configure(ChannelPipeline pipeline);
-}